requirements.txt
youtube_transcript_api
streamlit
groq
python-dotenv
pathlib
pyperclip
beautifulsoup4
requests
